+++ Pipfile
@@ -23,16 +23,15 @@ nest-asyncio = "*"
 
 [dev-packages]
 ruff = "*"
+pre-commit = "*"
 
 [requires]
 python_version = "3.12"
 
 [scripts]
-start-client = "python ./src/util/start_client.py"
 start-house-node = "python ./src/util/start_house_node.py"
-start-server = "python ./src/util/start_server.py"
 start-mock-node = "python ./src/util/start_mock_node.py"
-start-node = "python ./src/util/start_node.py"
-start-charts = "python -m src.live_charting"
-init-db = "python -m src.sqlite"
-test = "pytest"
+format = "ruff format"
+#start-node = "python ./src/util/start_node.py"
+#start-charts = "python -m src.live_charting"
+#init-db = "python -m src.sqlite"
