+++ src/mock_node/slider_values.txt
@@ -1,24 +1,24 @@
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-30
-24
-19
-30
-30
+0
+2
+1
+2
+1
+1
+1
+2
+4
+4
+3
+1
+1
+8
+7
+2
+2
+3
+4
+6
+12
 10
-21
-30
-30
+4
+3
