+++ src/openadr_node/__init__.py
@@ -1,3 +1,9 @@
-from src.legacy.async_event_bus import dispatcher
+# openadr_node/__init__.py
 
-__all__ = ['dispatcher']
+from src.openadr_node.adr_logger.logger import logger
+from src.openadr_node.node_manager import NodeManager
+from src.openadr_node.virtual_end_node import VirtualEndNode
+from src.openadr_node.virtual_top_node import VirtualTopNode
+from src.openadr_node.adr_base_config import AdrBaseConfig
+
+__all__ = ['NodeManager', 'VirtualEndNode', 'VirtualTopNode', 'AdrBaseConfig', 'logger']
