+++ src/openadr_node/adr_logger/__init__.py
@@ -1,0 +1,3 @@
+from src.openadr_node.adr_logger.logger import logger
+
+__all__ = ['logger']
