+++ src/openadr_node/adr_logger/logger.py
@@ -1,7 +1,4 @@
-# decorator/__init__.py
 import logging
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-__all__ = ['logger']
