+++ src/openadr_node/decorator/signal_connector.py
@@ -1,14 +1,13 @@
 from functools import wraps
 from pydispatch import dispatcher
-import logging
 
-logger = logging.getLogger(__name__)
+from src.openadr_node import logger
 
 
-class SignalSender:
+class SignalConnector:
   def __init__(self, signal=None, sender=None):
     """
-    Initialize the SendDispatcher decorator.
+    Initialize the ConnectDispatcher decorator.
 
     Args:
         signal (str, optional): Custom signal to use. If not provided, the method name will be used.
@@ -16,25 +15,6 @@ class SignalSender:
     """
     self._custom_signal = signal
     self._custom_sender = sender
-    self._ready_dispatched = False
-
-    dispatcher.connect(self.on_ready, signal='on_ready', sender=dispatcher.Any)
-
-  def on_ready(self, sender):
-    """
-    This method is called when the "on_ready" event is dispatched.
-    It will enable the dispatching of the `register_dispatcher` event.
-    """
-    logger.info('on_ready event received. Ready to register dispatchers.')
-    if not self._ready_dispatched:
-      # Dispatch the "register_dispatcher" event only once when "on_ready" is received
-      dispatcher.send(
-        signal='register_dispatcher',
-        sender=self._custom_sender,
-        data=self._custom_signal,
-      )
-      logger.info(f"Dispatched 'register_dispatcher' signal from {self._custom_sender}")
-      self._ready_dispatched = True
 
   def decorate(self, func):
     """
@@ -46,20 +26,13 @@ class SignalSender:
 
     @wraps(func)
     def wrapper(*args, **kwargs):
-      if not self._ready_dispatched:
-        logger.warning(
-          'on_ready event has not been dispatched. Aborting function call.'
-        )
-        return None
+      # Call the original method
+      return func(*args, **kwargs)
 
-      result = func(*args, **kwargs)
-
-      dispatcher.send(
-        signal=self._custom_signal, sender=self._custom_sender, data=result
-      )
-      logger.info(f"Dispatched signal '{self._custom_signal}' with result: {result}")
-
-      return result
+    # Attach signal and sender information for connection later
+    wrapper._signal = self._custom_signal or func.__name__
+    wrapper._sender = self._custom_sender
+    wrapper._original_func = func
 
     return wrapper
 
@@ -78,14 +51,17 @@ class SignalSender:
         instance: The instance of the class
     """
     for name, method in vars(instance.__class__).items():
+      # Check if the method was decorated
       if hasattr(method, '_signal') and hasattr(method, '_original_func'):
         # Create a bound method for the instance
         bound_method = method.__get__(instance, instance.__class__)
 
         # Connect the bound method to the dispatcher
+
         dispatcher.connect(
           receiver=bound_method, signal=method._signal, sender=method._sender
         )
+        # dispatcher.send(signal="register_dispatcher", sender=method._sender, data=method._signal)
         logger.info(
           f'Connected {name} to signal: {method._signal} with sender: {method._sender}'
         )
