+++ src/openadr_node/decorator/signal_sender.py
@@ -2,13 +2,11 @@ from functools import wraps
 from pydispatch import dispatcher
 import logging
 
-logger = logging.getLogger(__name__)
 
-
-class SignalConnector:
+class SignalSender:
   def __init__(self, signal=None, sender=None):
     """
-    Initialize the ConnectDispatcher decorator.
+    Initialize the SendDispatcher decorator.
 
     Args:
         signal (str, optional): Custom signal to use. If not provided, the method name will be used.
@@ -16,6 +14,27 @@ class SignalConnector:
     """
     self._custom_signal = signal
     self._custom_sender = sender
+    self._ready_dispatched = False
+
+    dispatcher.connect(self.on_ready, signal='on_ready', sender=dispatcher.Any)
+
+  def on_ready(self, sender):
+    """
+    This method is called when the "on_ready" event is dispatched.
+    It will enable the dispatching of the `register_dispatcher` event.
+    """
+    logging.info('on_ready event received. Ready to register dispatchers.')
+    if not self._ready_dispatched:
+      # Dispatch the "register_dispatcher" event only once when "on_ready" is received
+      dispatcher.send(
+        signal='register_dispatcher',
+        sender=self._custom_sender,
+        data=self._custom_signal,
+      )
+      logging.info(
+        f"Dispatched 'register_dispatcher' signal from {self._custom_sender}"
+      )
+      self._ready_dispatched = True
 
   def decorate(self, func):
     """
@@ -27,13 +46,20 @@ class SignalConnector:
 
     @wraps(func)
     def wrapper(*args, **kwargs):
-      # Call the original method
-      return func(*args, **kwargs)
+      if not self._ready_dispatched:
+        logging.warning(
+          'on_ready event has not been dispatched. Aborting function call.'
+        )
+        return None
 
-    # Attach signal and sender information for connection later
-    wrapper._signal = self._custom_signal or func.__name__
-    wrapper._sender = self._custom_sender
-    wrapper._original_func = func
+      result = func(*args, **kwargs)
+
+      dispatcher.send(
+        signal=self._custom_signal, sender=self._custom_sender, data=result
+      )
+      logging.info(f"Dispatched signal '{self._custom_signal}' with result: {result}")
+
+      return result
 
     return wrapper
 
@@ -52,19 +78,14 @@ class SignalConnector:
         instance: The instance of the class
     """
     for name, method in vars(instance.__class__).items():
-      # Check if the method was decorated
       if hasattr(method, '_signal') and hasattr(method, '_original_func'):
         # Create a bound method for the instance
         bound_method = method.__get__(instance, instance.__class__)
 
         # Connect the bound method to the dispatcher
-        print('receiver', bound_method)
-        print('method._signal', method._signal)
-        print('method._sender', method._sender)
         dispatcher.connect(
           receiver=bound_method, signal=method._signal, sender=method._sender
         )
-        # dispatcher.send(signal="register_dispatcher", sender=method._sender, data=method._signal)
-        logger.info(
+        logging.info(
           f'Connected {name} to signal: {method._signal} with sender: {method._sender}'
         )
